requirements.txt
beautifulsoup4==4.4.1
Cython==0.23.4
Django==1.9
django-bootstrap3==6.2.2
docutils==0.12
Kivy-Garden==0.1.1
mysqlclient==1.3.7
numpy==1.10.2
pandas==0.17.1
Pygments==2.0.2
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
virtualenv==13.1.2
virtualenvwrapper-win==1.2.1
wheel==0.26.0
